move download test in nightly
