+++ openmetadata-ui/src/main/resources/ui/playwright/utils/ContextCenterUtil.ts
@@ -10,11 +10,174 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-import { APIRequestContext, Browser, expect, Page } from '@playwright/test';
+import {
+  APIRequestContext,
+  Browser,
+  expect,
+  Locator,
+  Page,
+  Response,
+} from '@playwright/test';
 import { PolicyRulesType } from '../support/access-control/PoliciesClass';
 import { UserClass } from '../support/user/UserClass';
 import { uuid } from './common';
 import { waitForAllLoadersToDisappear } from './entity';
+
+// ─── Document types ───────────────────────────────────────────────────────────
+
+export interface ContextCenterDocument {
+  id: string;
+  name: string;
+  displayName?: string;
+}
+
+export interface ContextCenterFolder {
+  id: string;
+  name: string;
+  displayName?: string;
+  fullyQualifiedName?: string;
+}
+
+export interface BulkOperationResult {
+  numberOfRowsPassed?: number;
+  numberOfRowsFailed?: number;
+}
+
+export interface BulkIdsRequest {
+  ids?: string[];
+}
+
+interface UploadMultipart {
+  file: {
+    name: string;
+    mimeType: string;
+    buffer: Buffer;
+  };
+  folder?: string;
+}
+
+interface CapturedDownload {
+  download: string;
+  href: string;
+}
+
+interface DownloadCaptureWindow extends Window {
+  __contextCenterCapturedDownloads?: CapturedDownload[];
+  __contextCenterOriginalAnchorClick?: HTMLAnchorElement['click'];
+}
+
+// ─── Document helpers ─────────────────────────────────────────────────────────
+
+export const parseResponseJson = <T>(body: string): T => JSON.parse(body) as T;
+
+export const uploadDocument = async (
+  apiContext: APIRequestContext,
+  name: string,
+  buffer: Buffer,
+  folderFqn?: string
+): Promise<ContextCenterDocument> => {
+  const multipart: UploadMultipart = {
+    file: {
+      name,
+      mimeType: 'text/plain',
+      buffer,
+    },
+  };
+
+  if (folderFqn) {
+    multipart.folder = folderFqn;
+  }
+
+  const response = await apiContext.post(
+    '/api/v1/contextCenter/drive/files/upload',
+    { multipart }
+  );
+  const body = await response.text();
+  expect(response.status(), body).toBe(201);
+
+  return parseResponseJson<ContextCenterDocument>(body);
+};
+
+export const getDocumentRowByName = (page: Page, fileName: string): Locator =>
+  page
+    .getByTestId('documents-view')
+    .locator('[data-testid^="document-row-"]')
+    .filter({ hasText: fileName });
+
+export const selectDocumentByName = async (page: Page, fileName: string) => {
+  const row = getDocumentRowByName(page, fileName);
+  await expect(row).toBeVisible();
+  await row.scrollIntoViewIfNeeded();
+  await row.getByTestId('document-checkbox').click();
+};
+
+export const expectSelectedCount = async (page: Page, count: number) => {
+  await expect(
+    page.getByText(`${count} selected`, { exact: true })
+  ).toBeVisible();
+};
+
+export const expectBulkIdsRequest = (
+  postData: string | null,
+  expectedIds: string[]
+) => {
+  expect(postData).toBeTruthy();
+
+  const request = parseResponseJson<BulkIdsRequest>(postData ?? '{}');
+
+  expect(new Set(request.ids)).toEqual(new Set(expectedIds));
+};
+
+export const responseMatchesRequestPath = (
+  response: Response,
+  expectedPath: string
+): boolean => {
+  let request = response.request();
+
+  while (request) {
+    if (request.url().includes(expectedPath)) {
+      return true;
+    }
+
+    request = request.redirectedFrom();
+  }
+
+  return false;
+};
+
+export const installDownloadCapture = async (page: Page) => {
+  await page.evaluate(() => {
+    const captureWindow = window as DownloadCaptureWindow;
+    captureWindow.__contextCenterCapturedDownloads = [];
+
+    if (captureWindow.__contextCenterOriginalAnchorClick) {
+      return;
+    }
+
+    captureWindow.__contextCenterOriginalAnchorClick =
+      HTMLAnchorElement.prototype.click;
+    HTMLAnchorElement.prototype.click = function (this: HTMLAnchorElement) {
+      captureWindow.__contextCenterCapturedDownloads?.push({
+        download: this.download,
+        href: this.href,
+      });
+      captureWindow.__contextCenterOriginalAnchorClick?.call(this);
+    };
+  });
+};
+
+export const expectCapturedDownload = async (page: Page, fileName: string) => {
+  const capturedDownload = await page.waitForFunction((expectedFileName) => {
+    const captureWindow = window as DownloadCaptureWindow;
+
+    return captureWindow.__contextCenterCapturedDownloads?.find(
+      (download) => download.download === expectedFileName
+    );
+  }, fileName);
+  const download = (await capturedDownload.jsonValue()) as CapturedDownload;
+
+  expect(download.href.startsWith('blob:')).toBe(true);
+};
 
 export const DASHBOARD_URL = '/context-center/dashboard';
 export const ARTICLES_URL = '/context-center/articles';
